+++ src/main/java/me/pajic/rearm/item/NetheriteCrossbowItem.java
@@ -2,6 +2,9 @@ package me.pajic.rearm.item;
 
 import me.pajic.rearm.Main;
 import net.minecraft.core.component.DataComponents;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.flag.FeatureFlagSet;
 import net.minecraft.world.item.CrossbowItem;
 import net.minecraft.world.item.ItemStack;
@@ -15,8 +18,10 @@ public class NetheriteCrossbowItem extends CrossbowItem {
         super(
                 new Properties().fireResistant().stacksTo(1).durability(704)
                         .component(DataComponents.CHARGED_PROJECTILES, ChargedProjectiles.EMPTY)
-                        //? if 1.21.7
-                        /*.repairable(Items.NETHERITE_INGOT).enchantable(15)*/
+                        //? if 1.21.7 {
+                        /*.repairable(Items.NETHERITE_INGOT).enchantable(15)
+                        .setId(ResourceKey.create(Registries.ITEM, ResourceLocation.fromNamespaceAndPath(Main.MOD_ID, "netherite_crossbow")))
+                         *///?}
         );
     }