+++ versions/1.21.1/src/main/java/me/pajic/rearm/projectile/ThrownAxeRenderer.java
@@ -44,6 +44,7 @@ public class ThrownAxeRenderer extends EntityRenderer<ThrownAxe> {
         poseStack.popPose();
     }
 
+    @SuppressWarnings("deprecation")
     @Override
     public @NotNull ResourceLocation getTextureLocation(@NotNull ThrownAxe entity) {
         return TextureAtlas.LOCATION_BLOCKS;